MediaDrm#mNativeContext is a long, not int.
Therefore, we must use SetLongField and not SetIntField.

+++ media/jni/android_media_MediaDrm.cpp
@@ -491,7 +491,7 @@ static sp<JDrm> setDrm(
     if (old != NULL) {
         old->decStrong(thiz);
     }
-    env->SetIntField(thiz, gFields.context, (int)drm.get());
+    env->SetLongField(thiz, gFields.context, (int)drm.get());
 
     return old;
 }
